refactor: Add loading element and styles for improved user experience during asynchronous operations
- Introduced a new loading.js file that contains functions to create a loading element and add necessary styles.
- Updated showLoading function to utilize the new loading element and styles, enhancing the visual feedback during loading states.
- Removed redundant loading element creation code from showLoading for better maintainability.
